moved signaleventdefinition
